Rename : 데이터베이스 관련 클래스 이름 변경
특정 데이터베이스 이름이 아닌, 추상화된 데이터베이스 이름을 사용하였다.
왜냐하면, DI를 통해, 또는 데이터 베이스 Driver의 변경으로 데이터베이스를 쉽게 바꿀 수 있기 때문이다.
